Fix review findings: scope Carbon overrides, restore a11y, fix modal bugs
app.scss:
- Drop !important from tile/input/select/tag/label overrides so inline
  styles, Tailwind utilities, and Carbon state styling (disabled,
  invalid, size variants) win again; the sheet compiles after Carbon so
  source order already beats Carbon's base rules
- Scope select-input styling to exclude inline (Pagination) and disabled
  selects, and stop forcing height so size="sm" works
- Exclude dismissible filter tags from the tag geometry override
- Restore accessible 2px focus outlines (var(--cds-focus)) instead of
  outline: none on Carbon and native form controls
- Reserve icon space on invalid/warning text inputs
- Reset Carbon's footer-button block padding and baseline alignment so
  the 2.25rem footer buttons center their labels
- Remove dead .cds--modal transition line and the superseded dark-mode
  select fix blocks

Modal components:
- Extract shared ThemedModalHeader + modal/paper styles used by the
  cloud cost drilldown and external cost details modals
- Add dialog semantics (role, aria-modal, aria-labelledby) and an
  accessible close-button label; drop leftover MUI Modal title props
  that rendered browser tooltips
- Theme the recharts tooltip so it stays readable in dark mode
- Fix null-unsafe regex match, dead 404 check, and always-truthy chart
  guard in cloudCostDetails

Misc:
- Add type="text" to the untyped tag/filter inputs so the native focus
  rule matches them
- Position the filter autocomplete listbox absolutely so it overlays
  instead of pushing the layout
