+++ app/components/legacy/externalCosts/externalCostDetailModal.jsx
@@ -6,42 +6,16 @@ import {
   Table,
   TableBody,
 } from "@mui/material";
-
-const modalStyle = {
-  display: "flex",
-  alignItems: "center",
-  justifyContent: "center",
-};
+import {
+  ThemedModalHeader,
+  modalCenterStyle,
+  paperBaseStyle,
+} from "../themed-modal";
 
 const paperStyle = {
-  backgroundColor: "var(--cds-layer)",
-  color: "var(--cds-text-primary)",
-  border: "1px solid var(--cds-border-subtle)",
-  borderRadius: "8px",
-  boxShadow: "0 20px 25px -5px rgba(0, 0, 0, 0.15), 0 10px 10px -5px rgba(0, 0, 0, 0.04)",
-  padding: "24px",
+  ...paperBaseStyle,
   maxWidth: "600px",
-  width: "90%",
   maxHeight: "80vh",
-  overflowY: "auto",
-  outline: "none",
-};
-
-const headerStyle = {
-  display: "flex",
-  justifyContent: "space-between",
-  alignItems: "center",
-  marginBottom: "16px",
-  borderBottom: "1px solid var(--cds-border-subtle)",
-  paddingBottom: "12px",
-};
-
-const titleStyle = {
-  margin: 0,
-  fontSize: "1.25rem",
-  fontWeight: 600,
-  color: "var(--cds-text-primary)",
-  fontFamily: '"IBM Plex Sans", sans-serif',
 };
 
 const cellKeyStyle = {
@@ -60,39 +34,18 @@ const cellValStyle = {
 // for now, we can assume that the "Name" is resourceType
 export const ExternalCostDetails = ({ row, onClose }) => (
   <div>
-    <Modal
-      open={true}
-      onClose={onClose}
-      title={row.resource_type}
-      style={modalStyle}
-    >
-      <Paper style={paperStyle}>
-        <div style={headerStyle}>
-          <h3 style={titleStyle}>
-            {row.resource_type || "External Cost Details"}
-          </h3>
-          <button
-            onClick={onClose}
-            style={{
-              background: "none",
-              border: "none",
-              color: "var(--cds-text-primary)",
-              cursor: "pointer",
-              fontSize: "1.5rem",
-              lineHeight: 1,
-              padding: "4px 8px",
-              borderRadius: "4px",
-            }}
-            onMouseEnter={(e) => {
-              e.currentTarget.style.backgroundColor = "var(--cds-layer-hover)";
-            }}
-            onMouseLeave={(e) => {
-              e.currentTarget.style.backgroundColor = "transparent";
-            }}
-          >
-            &times;
-          </button>
-        </div>
+    <Modal open={true} onClose={onClose} style={modalCenterStyle}>
+      <Paper
+        style={paperStyle}
+        role="dialog"
+        aria-modal="true"
+        aria-labelledby="external-cost-details-title"
+      >
+        <ThemedModalHeader
+          title={row.resource_type || "External Cost Details"}
+          titleId="external-cost-details-title"
+          onClose={onClose}
+        />
         <TableContainer>
           <Table>
             <TableBody>
